♻️ Refactor Checks Into Modular Files
