+++ app/projects/repository_standards/config/repository_compliance_config.py
@@ -1,0 +1,152 @@
+from typing import List, Optional
+from app.projects.repository_standards.models.repository_compliance import (
+    RepositoryComplianceCheck,
+)
+from app.projects.repository_standards.repositories.asset_repository import (
+    RepositoryView,
+)
+
+PASS = "pass"
+FAIL = "fail"
+
+
+def get_secret_scanning_enabled_check(
+    repository: RepositoryView, required: bool = True
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Secret Scanning Enabled",
+        status=PASS
+        if repository.data.security_and_analysis.secret_scanning_status == "enabled"
+        else FAIL,
+        required=required,
+        description="Imporves organisational security by scanning and reporting secrets.",
+    )
+
+
+def get_secret_scanning_push_protection_enabled_check(
+    repository: RepositoryView, required: bool = True
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Secret Scanning Push Protection Enabled",
+        status=PASS
+        if repository.data.security_and_analysis.push_protection_status == "enabled"
+        else FAIL,
+        required=required,
+        description="Prevents secrets from being pushed to the repository.",
+    )
+
+
+def get_branch_protection_enforced_for_admins_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch Protection Enforced For Admins",
+        status=PASS
+        if repository.data.default_branch_protection.enforce_admins
+        else FAIL,
+        required=required,
+        description="Prevents admins from bypassing branch protection.",
+    )
+
+
+def get_default_branch_protection_requires_signed_commits_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch Protection Requires Signed Commits",
+        status=PASS
+        if repository.data.default_branch_protection.required_signatures
+        else FAIL,
+        required=required,
+        description="Signed commits ensure that the commit author is verified, preventing impersonations.",
+    )
+
+
+def get_default_branch_protection_requires_code_owner_reviews_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch Protection Requires Code Owner Reviews",
+        status=PASS
+        if repository.data.default_branch_protection.require_code_owner_reviews
+        else FAIL,
+        required=required,
+        description="Useful for delegating reviews of parts of the codebase to specific people.",
+    )
+
+
+def get_default_branch_pull_requests_dismiss_stale_reviews_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch Pull Request Dismiss Stale Reviews",
+        status=PASS
+        if repository.data.default_branch_protection.dismiss_stale_reviews
+        else FAIL,
+        required=required,
+        description="Ensures that the latest changes are reviewed before merging.",
+    )
+
+
+def get_default_branch_protection_requires_atleast_one_review_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch Pull Request Requires Atleast One Review",
+        status=PASS
+        if repository.data.default_branch_protection.required_approving_review_count
+        or 0 >= 1
+        else FAIL,
+        required=required,
+        description="Ensures that at least one person has reviewed the changes before merging.",
+    )
+
+
+def get_has_authorative_owner_check(
+    authorative_owner: Optional[str], required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Has an Authorative Owner",
+        status=PASS if authorative_owner else FAIL,
+        required=required,
+        description="Prevents orphaned repositories by having an easily identifiable owner.",
+    )
+
+
+def get_licence_is_mit_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="License is MIT",
+        status=PASS if repository.data.basic.license == "mit" else FAIL,
+        required=required,
+        description="MIT License is a permissive license that allows for reuse of the codebase.",
+    )
+
+
+def get_default_branch_is_main_check(
+    repository: RepositoryView, required: bool = False
+) -> RepositoryComplianceCheck:
+    return RepositoryComplianceCheck(
+        name="Default Branch is main",
+        status=PASS if repository.data.basic.default_branch_name == "main" else FAIL,
+        required=required,
+        description="main is a more inclusive and modern term for the default branch.",
+    )
+
+
+def get_all_compliance_checks(
+    repository: RepositoryView, authorative_owner: Optional[str]
+) -> List[RepositoryComplianceCheck]:
+    return [
+        get_secret_scanning_enabled_check(repository),
+        get_secret_scanning_push_protection_enabled_check(repository),
+        get_branch_protection_enforced_for_admins_check(repository),
+        get_default_branch_protection_requires_signed_commits_check(repository),
+        get_default_branch_protection_requires_code_owner_reviews_check(repository),
+        get_default_branch_pull_requests_dismiss_stale_reviews_check(repository),
+        get_default_branch_protection_requires_atleast_one_review_check(repository),
+        get_has_authorative_owner_check(authorative_owner),
+        get_licence_is_mit_check(repository),
+        get_default_branch_is_main_check(repository),
+    ]
